+++ cupy/random/_generator_api.pyx
